[LABY] Fixes after the migration
