+++ emma-examples-labyrinth/src/main/scala/org/emmalanguage/cli/LabyrinthExamplesRunner.scala
