+++ emma-labyrinth/src/test/java/org/emmalanguage/labyrinth/inputgen/ClickCountDiffsInputGen.java
@@ -1,4 +1,20 @@
-package inputgen;
+/*
+ * Copyright © 2014 TU Berlin (emma@dima.tu-berlin.de)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.emmalanguage.labyrinth.inputgen;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -77,7 +93,7 @@ public class ClickCountDiffsInputGen {
         }
     }
 
-    static String readFile(String path, Charset encoding)
+    private static String readFile(String path, Charset encoding)
             throws IOException
     {
         byte[] encoded = Files.readAllBytes(Paths.get(path));
