+++ emma-labyrinth/src/test/java/org/emmalanguage/labyrinth/jobs/ClickCountDiffs.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package jobs;
+package org.emmalanguage.labyrinth.jobs;
 
 import org.emmalanguage.labyrinth.ElementOrEvent;
 import org.emmalanguage.labyrinth.LabyNode;
