+++ emma-labyrinth/src/test/java/org/emmalanguage/labyrinth/jobsold/ConnectedComponentsMB.java
@@ -14,36 +14,36 @@
  * limitations under the License.
  */
 
-package jobsold;
+package org.emmalanguage.labyrinth.jobsold;
 
 import org.emmalanguage.labyrinth.BagOperatorHost;
 import org.emmalanguage.labyrinth.CondOutputSelector;
+import org.emmalanguage.labyrinth.ElementOrEvent;
+import org.emmalanguage.labyrinth.MutableBagCC;
+import org.emmalanguage.labyrinth.PhiNode;
 import org.emmalanguage.labyrinth.operators.Bagify;
 import org.emmalanguage.labyrinth.operators.DistinctInt;
-import org.emmalanguage.labyrinth.operators.IdMap;
-import org.emmalanguage.labyrinth.partitioners.Always0;
-import org.emmalanguage.labyrinth.operators.*;
-import org.emmalanguage.labyrinth.partitioners.TupleIntIntBy0;
-import org.emmalanguage.labyrinth.partitioners.FlinkPartitioner;
-import org.emmalanguage.labyrinth.partitioners.Forward;
-import org.emmalanguage.labyrinth.util.LogicalInputIdFiller;
-import org.emmalanguage.labyrinth.util.TupleIntInt;
-import org.emmalanguage.labyrinth.util.Unit;
-import org.emmalanguage.labyrinth.util.Util;
-import org.emmalanguage.labyrinth.CFLConfig;
-import org.emmalanguage.labyrinth.ElementOrEvent;
-import org.emmalanguage.labyrinth.KickoffSource;
-import org.emmalanguage.labyrinth.PhiNode;
-import org.emmalanguage.labyrinth.operators.AssertBagEquals;
-import org.emmalanguage.labyrinth.operators.ConditionNode;
-import org.emmalanguage.labyrinth.operators.FlatMap;
 import org.emmalanguage.labyrinth.operators.GraphSinkTupleIntInt;
 import org.emmalanguage.labyrinth.operators.GroupBy0Min1TupleIntInt;
+import org.emmalanguage.labyrinth.operators.IdMap;
 import org.emmalanguage.labyrinth.operators.JoinTupleIntInt;
-import org.emmalanguage.labyrinth.operators.NonEmptyCombiner;
 import org.emmalanguage.labyrinth.operators.Or;
-import org.emmalanguage.labyrinth.operators.UpdateJoinTupleIntInt;
+import org.emmalanguage.labyrinth.partitioners.Always0;
+import org.emmalanguage.labyrinth.partitioners.TupleIntIntBy0;
+import org.emmalanguage.labyrinth.util.LogicalInputIdFiller;
+import org.emmalanguage.labyrinth.util.Unit;
+import org.emmalanguage.labyrinth.CFLConfig;
+import org.emmalanguage.labyrinth.KickoffSource;
+import org.emmalanguage.labyrinth.operators.AssertBagEquals;
+import org.emmalanguage.labyrinth.util.TupleIntInt;
+import org.emmalanguage.labyrinth.util.Util;
+import org.emmalanguage.labyrinth.operators.ConditionNode;
+import org.emmalanguage.labyrinth.operators.FlatMap;
+import org.emmalanguage.labyrinth.operators.NonEmptyCombiner;
+import org.emmalanguage.labyrinth.partitioners.FlinkPartitioner;
+import org.emmalanguage.labyrinth.partitioners.Forward;
 import org.apache.flink.api.common.ExecutionConfig;
+import org.apache.flink.api.common.functions.FilterFunction;
 import org.apache.flink.api.common.typeinfo.TypeHint;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
@@ -63,49 +63,52 @@ import org.slf4j.LoggerFactory;
 // edges = // read graph (each edge has to be present twice: (u,v) and (v,u))
 // vertices0 = edges.map((x,y) => x)
 // vertices = vertices0.distinct
-// labels = vertices.map(x => (x,x))
+// labels = toMutable(vertices.map(x => (x,x)), (x,y) => x)
 // updates = labels
 // do {
 //     // BB 1
 //     msgs = updates.join(edges).on(0).equalTo(0).with( ((vid, label), (u,v)) => (v,label) )
 //     minMsgs = msgs.groupBy(0).min(1)
-//     updates = labels.join(minMsgs).with( ((vid, label), (target, msg)) => if (msg < label) out.collect((target, msg)) )
-//     labels = labels.update(updates)
+//     updates = join(labels, minMsgs, (x,y) => x)
+// 			.flatMap((vid, label), (target, msg)) => if (msg < label) out.collect((target, msg)) )
+//     update(labels, updates)
 // } while (!updates.isEmpty)
 // // BB 2
-// labels.write
+// toBag(labels).write
 // assertBagEquals(labels_2, ...)
 
 
 // SSA
 // // BB 0
 // edges = // read graph (each edge has to be present twice: (u,v) and (v,u))
-// vertices0 = edges.map((x,y) => x)
-// vertices = vertices0.distinct
+// verticesMult = edges.map((x,y) => x)
+// vertices = verticesMult.distinct
 // labels_0 = vertices.map(x => (x,x))
+// mb = toMutable(labels_0)
 // updates_0 = labels_0
 // do {
 //     // BB 1
-//     labels_1 = phi(labels_0, labels_2)
 //     updates_1 = phi(updates_0, updates_2)
 //     msgs = edges.join(updates_1).on(0).equalTo(0).with( ((u,v), (vid, label)) => (v,label) )
 //     minMsgs = msgs.groupBy(0).min(1)
-//     updates_2 = labels_1.join(minMsgs).on(0).equalTo(0).with( ((vid, label), (target, msg)) => if (msg < label) out.collect((target, msg)) )
-//     labels_2 = labels_1.update(updates_2)
+//     j = join(mb, minMsgs) // a kov. sor is beolvasztva most ebbe
+//     updates_2 = j.flatMap( ((vid, label), (target, msg)) => if (msg < label) out.collect((target, msg)) )
+//     update(mb, updates_2)
 //     nonEmpty = updates_2.nonEmpty
 // } while (nonEmpty)
 // // BB 2
-// labels_2.write
-// assertBagEquals(labels_2, ...)
+// result = toBag(mb)
+// result.write
+// assertBagEquals(result, ...)
 
 
-public class ConnectedComponents {
+public class ConnectedComponentsMB {
 
-	private static final Logger LOG = LoggerFactory.getLogger(ConnectedComponents.class);
+	private static final Logger LOG = LoggerFactory.getLogger(ConnectedComponentsMB.class);
 
 	private static TypeSerializer<Integer> integerSer = TypeInformation.of(Integer.class).createSerializer(new ExecutionConfig());
 	private static TypeSerializer<Boolean> booleanSer = TypeInformation.of(Boolean.class).createSerializer(new ExecutionConfig());
-	private static TypeSerializer<TupleIntInt> tupleIntIntSer = new TupleIntInt.TupleIntIntSerializer();
+	public static TypeSerializer<TupleIntInt> tupleIntIntSer = new TupleIntInt.TupleIntIntSerializer();
 
 	public static void main(String[] args) throws Exception {
 		StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
@@ -119,7 +122,7 @@ public class ConnectedComponents {
 //		cfg.setLong("taskmanager.network.numberOfBuffers", 32768); //16384
 //		StreamExecutionEnvironment env = StreamExecutionEnvironment.createLocalEnvironment(20, cfg); //20 // amugy sokszor 9-nel jon elo hiba, mivel ekkor mar nehol nem kap minden instance
 
-		//env.getConfig().setParallelism(1);
+		//env.getConfig().setParallelism(1); /////////
 
 		int para = env.getParallelism();
 
@@ -177,9 +180,35 @@ public class ConnectedComponents {
 			}
 		}, 0, 2, integerSer)
 			.addInput(0,0,true,1)
-			.out(0,0,true, new Forward<>(para)))
+			.out(0,0,true, new TupleIntIntBy0(para)))
 				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
 				.setConnectionType(new FlinkPartitioner<>());
+
+
+
+		// ---------------- MutableBag ----------------
+
+		IterativeStream<ElementOrEvent<TupleIntInt>> mbIt = labels_0
+				.map(new LogicalInputIdFiller<>(0))
+				.iterate(1000000000);
+
+		DataStream<ElementOrEvent<TupleIntInt>> mbIt0 = mbIt.bt("MutableBagCC", Util.tpe(),
+				new MutableBagCC(15)
+			.addInput(0, 0, true, 2)
+			.addInput(1, 1, true, 7)
+			.addInput(2, 1, true, 15)
+			.out(0,1,true, new TupleIntIntBy0(para)) // To update
+			.out(1,1,true, new Forward<>(para)) // to nonEmpty
+			.out(2,1,false, new Forward<>(para)) // back-edge to updates_1 phi-node
+			.out(3, 2, true, outFile == null ? new Always0<>(1) : new Forward<>(para)) // output of toBag
+		)
+				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
+				.setConnectionType(new FlinkPartitioner<>());
+
+		SplitStream<ElementOrEvent<TupleIntInt>> mbSplit = mbIt0.split(new CondOutputSelector<>());
+
+
+
 
 		DataStream<ElementOrEvent<TupleIntInt>> updates_0 = labels_0
 				.bt("updates_0", Util.tpe(), new BagOperatorHost<>(new IdMap<TupleIntInt>(), 0, 3, tupleIntIntSer)
@@ -191,23 +220,12 @@ public class ConnectedComponents {
 
 		// --- Iteration starts here ---
 
-		IterativeStream<ElementOrEvent<TupleIntInt>> labelsIt = labels_0.map(new LogicalInputIdFiller<>(0)).iterate(1000000000);
-
-		DataStream<ElementOrEvent<TupleIntInt>> labels_1 = labelsIt
-				.bt("labels_1", Util.tpe(), new PhiNode<TupleIntInt>(1, 4, tupleIntIntSer)
-				.addInput(0,0,false,2)
-				.addInput(1,1,false,9)
-				.out(0,1,true, new TupleIntIntBy0(para)) // (this goes to two places, but none of them is conditional)
-		)
-				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
-				.setConnectionType(new FlinkPartitioner<>());
-
 		IterativeStream<ElementOrEvent<TupleIntInt>> updatesIt = updates_0.map(new LogicalInputIdFiller<>(0)).iterate(1000000000);
 
 		DataStream<ElementOrEvent<TupleIntInt>> updates_1 = updatesIt
 				.bt("updates_1", Util.tpe(), new PhiNode<TupleIntInt>(1, 5, tupleIntIntSer)
 				.addInput(0,0,false,3)
-				.addInput(1,1,false,8)
+				.addInput(1,1,false,15)
 				.out(0,1,true, new TupleIntIntBy0(para))
 		)
 				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
@@ -217,7 +235,7 @@ public class ConnectedComponents {
 		DataStream<ElementOrEvent<TupleIntInt>> msgs = edges
 				.map(new LogicalInputIdFiller<>(0))
 				.union(updates_1.map(new LogicalInputIdFiller<>(1)))
-				.setConnectionType(new FlinkPartitioner<>()) // Meg lehetne nezni, hogy enelkul is mukodik-e
+				//.setConnectionType(new FlinkPartitioner<>()) // Ugy tunik, hogy enelkul is mukodik
 				.bt("msgs", Util.tpe(), new BagOperatorHost<>(new JoinTupleIntInt<TupleIntInt>(){
 					@Override
 					protected void udf(int b, TupleIntInt p) {
@@ -239,54 +257,22 @@ public class ConnectedComponents {
 				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
 				.setConnectionType(new FlinkPartitioner<>());
 
-		//     updates_2 = labels_1.join(minMsgs).on(0).equalTo(0).with( ((vid, label), (target, msg)) => if (msg < label) out.collect((target, msg)) )
-		SplitStream<ElementOrEvent<TupleIntInt>> updates_2 = labels_1
-				.map(new LogicalInputIdFiller<>(0))
-				.union(minMsgs.map(new LogicalInputIdFiller<>(1)))
-				.setConnectionType(new FlinkPartitioner<>())
-				.bt("updates_2", Util.tpe(), new BagOperatorHost<>(new JoinTupleIntInt<TupleIntInt>(){
+		mbIt.closeWith(minMsgs.map(new LogicalInputIdFiller<>(1))
+			.union(mbIt0.filter(new FilterFunction<ElementOrEvent<TupleIntInt>>() {
 					@Override
-					protected void udf(int b, TupleIntInt p) {
-						if (b > p.f1) {
-							out.collectElement(p);
-						}
+					public boolean filter(ElementOrEvent<TupleIntInt> tuple2ElementOrEvent) throws Exception {
+						return false;
 					}
-				}, 1, 8, tupleIntIntSer)
-								.addInput(0,1,true,4)
-								.addInput(1,1,true,7)
-								.out(0,1,true, new TupleIntIntBy0(para)) // To labels_2 update
-								.out(1,1,true, new Forward<>(para)) // to nonEmpty
-						        .out(2,1,false, new Forward<>(para)) // back-edge
-				)
-				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
-				.setConnectionType(new FlinkPartitioner<>())
-				.split(new CondOutputSelector<>());
+				}).returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
+			));
 
-		//     labels_2 = labels_1.update(updates_2)
-		SplitStream<ElementOrEvent<TupleIntInt>> labels_2 = labels_1
-				.map(new LogicalInputIdFiller<>(0))
-				.union(updates_2.select("0").map(new LogicalInputIdFiller<>(1)))
-				.setConnectionType(new FlinkPartitioner<>())
-				.bt("labels_2",Util.tpe(),new BagOperatorHost<>(new UpdateJoinTupleIntInt(),1,9, tupleIntIntSer)
-								.addInput(0,1,true,4) // labels_1
-								.addInput(1,1,true,8) // updates_2
-								.out(0,1,false, new Forward<>(para)) // back-edge
-								.out(1,2,false, outFile == null ? new Always0<>(1) : new Forward<>(para)) // out of the loop
-				)
-				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
-				.setConnectionType(new FlinkPartitioner<>())
-				.split(new CondOutputSelector<>());
+		mbIt.closeWith(mbSplit.select("0").map(new LogicalInputIdFiller<>(2)));
 
-		labelsIt.closeWith(labels_2.select("0").map(new LogicalInputIdFiller<>(1)));
-		updatesIt.closeWith(updates_2.select("2").map(new LogicalInputIdFiller<>(1)));
-
-		// exit cond
-
-		DataStream<ElementOrEvent<Boolean>> nonEmpty = updates_2.select("1")
+		DataStream<ElementOrEvent<Boolean>> nonEmpty = mbSplit.select("1")
 				.bt("nonEmpty",Util.tpe(),
 						new BagOperatorHost<>(
 								new NonEmptyCombiner<TupleIntInt>(), 1, 10, tupleIntIntSer)
-								.addInput(0, 1, true, 8)
+								.addInput(0, 1, true, 15)
 								.out(0,1,true, new Always0<>(1)))
 				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<Boolean>>(){}));
 
@@ -306,11 +292,23 @@ public class ConnectedComponents {
 								.addInput(0, 1, true, 16))
 				.setParallelism(1);
 
+		updatesIt.closeWith(mbSplit.select("2").map(new LogicalInputIdFiller<>(1))
+				.returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
+			.union(updates_1.filter(new FilterFunction<ElementOrEvent<TupleIntInt>>() {
+					@Override
+					public boolean filter(ElementOrEvent<TupleIntInt> tuple2ElementOrEvent) throws Exception {
+						return false;
+					}
+				}).returns(TypeInformation.of(new TypeHint<ElementOrEvent<TupleIntInt>>(){}))
+			));
+
 		// --- Iteration ends here ---
 
 
+		DataStream<ElementOrEvent<TupleIntInt>> result = mbSplit.select("3");
+
 		if (outFile == null) {
-			labels_2.select("1").bt("AssertBagEquals", Util.tpe(), new BagOperatorHost<>(new AssertBagEquals<>(
+			result.bt("AssertBagEquals", Util.tpe(), new BagOperatorHost<>(new AssertBagEquals<>(
 					TupleIntInt.of(0, 0),
 					TupleIntInt.of(1, 0),
 					TupleIntInt.of(2, 0),
@@ -319,12 +317,13 @@ public class ConnectedComponents {
 					TupleIntInt.of(5, 5),
 					TupleIntInt.of(6, 5),
 					TupleIntInt.of(7, 5)
-			), 2, 13, tupleIntIntSer)
-					.addInput(0, 1, false, 9)).setParallelism(1);
+					), 2, 13, tupleIntIntSer)
+							.addInput(0, 1, true, 15)).setParallelism(1);
 		} else {
-			labels_2.select("1").bt("GraphSink", Util.tpe(), new BagOperatorHost<>(new GraphSinkTupleIntInt(outFile), 2, 13, tupleIntIntSer)
-					.addInput(0,1,false,9));
+			result.bt("GraphSink", Util.tpe(), new BagOperatorHost<>(new GraphSinkTupleIntInt(outFile), 2, 13, tupleIntIntSer)
+							.addInput(0,1,true,15));
 		}
+
 
 		CFLConfig.getInstance().setNumToSubscribe();
